requirements.txt
alembic==1.6.5
click==8.0.1
ecdsa==0.19.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pyasn1==0.6.1
python-dateutil==2.8.1
python-dotenv==1.0.1
python-editor==1.0.4
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
